Remove cross-namespace ownership
To access encryption secrets, we create role and rolebinding
in the encryption secrets namespace. We used to add storagecluster
as the owner for these objects. K8s does not like having cross-
namespace owner references. If done so it garbage collects
all the objects owned by the StorageCluster when the garbage
collector restarts.

Signed-off-by: Piyush Nimbalkar <pnimbalkar@purestorage.com>