OPERATOR-519 Mount /var/lib/osd directly only for px > 2.9.1
oci-monitor 2.9.1 has a fix to allow /var/lib/osd mounted as a shared directory,
when the container runtime is cri-o.
Without the oci monitor change, mounting /var/lib/osd is not very useful as it
is not getting mounted as a shared mount. So keeping the old behavior for older
px releases, while mounting /var/lib/osd for 2.9.1 and above.

Signed-off-by: Piyush Nimbalkar <pnimbalkar@purestorage.com>