Add preserveFullCustomRegistry to not swallow part of image
This is to stop GetImageURN from swallowing part of the image tag when
setting a customImageRegistry with a / in it. Original behaviour is,
when a registry of `example.io/public` is used with an image of
`portworx/oci-monitor`, for the image to become
`emaple.io/public/oci-monitor`. Setting this to true will preserve the
full path and the image will be `example.io/public/portworx/oci-monitor`