OPERATOR-666 Set maxSurge and maxUnavailable in stork scheduler
In a 3 node cluster, the stork scheduler upgrades were getting stuck because of
the pod anti-affinity policy. The new 4th pod was going in a Pending state and
waiting for one of the existing 3 pods to go down. This was happening because
the default maxUnavailable is 25%. This was preventing the existing pod from
going down, and the pod anti-affinity stopped the new pod from coming up.

Signed-off-by: Piyush Nimbalkar <pnimbalkar@purestorage.com>