PWX-28442 Use quorum member flag to count storage nodes
- Earlier we were looking at the pools in the nodes enumerate response to
determine if a node is a storage node or not. This is not reliable as sometimes
during the intialization portworx does not populate the pools. This misleads the
operator to consider it as a storageless node.
- We recently added the NonQuorumMember flag in the nodes enumerate
response. This is more reliable and will always correctly tell whether a node is
storage or storageless.

Signed-off-by: Piyush Nimbalkar <pnimbalkar@purestorage.com>