PWX-37573: tighten the checks for avoiding unnecessary VM live migration
If the live-migration is in progress for a VM that we want to evict,
do not create another live migration for that VM in the same Reconcile() cycle.

Just before starting a live-migration, check the VMI one more time to verify
that the VMI is pointing to the same node as the pod being evicted.

These two extra checks reduce the window in which PX and operator may
try to live-migrate the same VM out of the same node. Also, it handles
any other unexpected live-migrations that might start.

Signed-off-by: Neelesh Thakur <neelesh.thakur@purestorage.com>