PWX-38103 Ignore unhealthy nodes if a healthy one is found
for the same k8s node. If one healthy node is found then there is a PX node
which is healthy on that k8s node. The unhealthy ones are present because they
were previously running on that k8s node before. Such nodes will either find a
new k8s node or will get auto-decommissioned in case of a storageless nodes.

Signed-off-by: Piyush Nimbalkar <pnimbalkar@purestorage.com>